code/requirements/requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
psycopg2-binary==2.9.9
sqlparse==0.5.3
django-environ==0.12.0